# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Intelligent scheduling system for UK personal trainers.",
        "title": "PT Scheduler API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "PT Scheduler Support",
            "email": "support@pt-scheduler.io"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/schedule-runs": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Runs the OR-Tools solver for the given week and returns proposed sessions for the coach to review.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Trigger a schedule run",
                "parameters": [
                    {
                        "description": "Week start date (Monday, YYYY-MM-DD)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.TriggerRunRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.ScheduleRun"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Infeasible schedule or constraint violation",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedule-runs/{runID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a schedule run and its proposed or confirmed sessions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Get a schedule run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule run UUID",
                        "name": "runID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.ScheduleRun"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedule-runs/{runID}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Coach approves the proposed schedule. All proposed sessions become confirmed. Notifications are enqueued.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Confirm a schedule run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule run UUID",
                        "name": "runID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.ScheduleRun"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Run is not pending confirmation",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedule-runs/{runID}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Coach rejects the proposed schedule. All proposed sessions are soft-deleted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Reject a schedule run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule run UUID",
                        "name": "runID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.ScheduleRun"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Run is not pending confirmation",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Coaches see all sessions across their clients. Clients see their own sessions. Filter by status with ?status=proposed|confirmed|cancelled|completed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "List sessions",
                "parameters": [
                    {
                        "enum": [
                            "proposed",
                            "confirmed",
                            "cancelled",
                            "completed"
                        ],
                        "type": "string",
                        "description": "Filter by session status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_scheduling.Session"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sessions/{sessionID}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Outside 24h window: cancels immediately and issues a credit. Inside 24h window: moves to pending_cancellation and notifies the coach to decide. Check within_24h_window in the response to know which path was taken.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Cancel a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session UUID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cancellation reason",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.CancelSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.CancelSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sessions/{sessionID}/cancel/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Coach confirms the client loses the session. No credit is issued. Only valid when session is in pending_cancellation status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scheduling"
                ],
                "summary": "Approve cancellation (session lost)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session UUID",
                        "name": "sessionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_scheduling.Session"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_danielgonzalez_pt-scheduler_internal_platform_httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_scheduling.CancelSessionRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
        "internal_scheduling.CancelSessionResponse": {
            "type": "object",
            "properties": {
                "credit": {
                    "$ref": "#/definitions/internal_scheduling.SessionCredit"
                },
                "message": {
                    "type": "string"
                },
                "session": {
                    "$ref": "#/definitions/internal_scheduling.Session"
                },
                "within_24h_window": {
                    "type": "boolean"
                }
            }
        },
        "internal_scheduling.ScheduleRun": {
            "type": "object",
            "properties": {
                "coach_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sessions": {
                    "description": "populated by service layer",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_scheduling.Session"
                    }
                },
                "solver_input": {},
                "solver_output": {},
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "week_start": {
                    "type": "string"
                }
            }
        },
...
        "internal_scheduling.TriggerRunRequest": {
            "type": "object",
            "required": [
                "week_start"
            ],
            "properties": {
                "week_start": {
                    "description": "\"YYYY-MM-DD\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer \" followed by your access token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}